+++ hydra-cardano-api/src/Cardano/Api/UTxO.hs
@@ -10,6 +10,8 @@ module Cardano.Api.UTxO where
 
 import Cardano.Api hiding (UTxO, toLedgerUTxO)
 import Cardano.Api qualified
+import Cardano.Api.Shelley (ReferenceScript (..))
+import Data.Bifunctor (second)
 import Data.Coerce (coerce)
 import Data.List qualified as List
 import Data.Map (Map)
@@ -84,8 +86,36 @@ min = UTxO . uncurry Map.singleton . Map.findMin . toMap
 
 -- * Type Conversions
 
-fromApi :: Cardano.Api.UTxO Era -> UTxO
-fromApi = coerce
+-- | Transforms a UTxO containing tx outs from any era into Babbage era.
+fromApi :: Cardano.Api.UTxO era -> UTxO
+fromApi (Cardano.Api.UTxO eraUTxO) =
+  let eraPairs = Map.toList eraUTxO
+      babbagePairs = second coerceOutputToEra <$> eraPairs
+   in fromPairs babbagePairs
+ where
+  coerceOutputToEra :: TxOut CtxUTxO era -> TxOut CtxUTxO Era
+  coerceOutputToEra (TxOut eraAddress eraValue eraDatum eraRefScript) =
+    TxOut
+      (coerceAddressToEra eraAddress)
+      (coerceValueToEra eraValue)
+      (coerceDatumToEra eraDatum)
+      (coerceRefScriptToEra eraRefScript)
+
+  coerceAddressToEra :: AddressInEra era -> AddressInEra Era
+  coerceAddressToEra (AddressInEra _ eraAddress) = anyAddressInShelleyBasedEra ShelleyBasedEraBabbage (toAddressAny eraAddress)
+
+  coerceValueToEra :: TxOutValue era -> TxOutValue Era
+  coerceValueToEra (TxOutAdaOnly _ eraLovelace) = lovelaceToTxOutValue BabbageEra eraLovelace
+  coerceValueToEra (TxOutValue _ value) = TxOutValue MaryEraOnwardsBabbage value
+
+  coerceDatumToEra :: TxOutDatum CtxUTxO era -> TxOutDatum CtxUTxO Era
+  coerceDatumToEra TxOutDatumNone = TxOutDatumNone
+  coerceDatumToEra (TxOutDatumHash _ hashScriptData) = TxOutDatumHash AlonzoEraOnwardsBabbage hashScriptData
+  coerceDatumToEra (TxOutDatumInline _ hashableScriptData) = TxOutDatumInline BabbageEraOnwardsBabbage hashableScriptData
+
+  coerceRefScriptToEra :: ReferenceScript era -> ReferenceScript Era
+  coerceRefScriptToEra ReferenceScriptNone = ReferenceScriptNone
+  coerceRefScriptToEra (ReferenceScript _ scriptInAnyLang) = ReferenceScript BabbageEraOnwardsBabbage scriptInAnyLang
 
 toApi :: UTxO -> Cardano.Api.UTxO Era
 toApi = coerce