+++ hydra-node/src/Hydra/Chain/Direct/ScriptRegistry.hs
@@ -157,7 +157,10 @@ registryUTxO scriptRegistry =
 -- Can throw at least 'NewScriptRegistryException' on failure.
 queryScriptRegistry ::
   (MonadIO m, MonadThrow m) =>
+  -- | cardano-node's network identifier.
+  -- A combination of network discriminant + magic number.
   NetworkId ->
+  -- | Filepath to the cardano-node's domain socket
   SocketPath ->
   TxId ->
   m ScriptRegistry